+++ src/clickgen/libs/colors.py
@@ -1,3 +1,7 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+
 class bcolors:
     BOLD = "\033[1m"